Refactor services into domain modules
