+++ magazyn/domain/__init__.py
@@ -1,0 +1,8 @@
+"""Domain layer modules for warehouse operations."""
+
+__all__ = [
+    "products",
+    "inventory",
+    "reports",
+    "invoice_import",
+]
